feat(tvos): make the channel column the EPG guide's focus hub
The guide's only escape from a deep horizontal scroll was walking every
programme cell back to the leading edge. The frozen channel column is now
the tvOS navigation hub: focus lands on it entering the guide, select plays
the channel, left reaches the category rail, right enters the programmes,
and Menu inside the grid snaps back to now and refocuses the column instead
of leaving the screen. The grid scrolls vertically to follow column focus.

Programme rows now realize their cells by exact timeline placement inside a
block-quantized window instead of a LazyHStack, whose estimated layout
misplaced rows realized deep into the 36-hour window by hours.
